Add remote_id to ingredients.
This allows to properly update them during imports. Currently, we were using the
code field for OFF products but e.g. USDA products have their own id
